+++ lib/src/database/commands/operation.dart
@@ -8,7 +8,6 @@ export 'aggregation_commands/wrapper/change_stream/change_stream_operation.dart'
 export 'aggregation_commands/wrapper/change_stream/change_stream_options.dart';
 
 export 'authentication_commands/sasl_continue_command/sasl_continue_command.dart';
-export 'authentication_commands/sasl_continue_command/sasl_continue_command.dart';
 export 'authentication_commands/sasl_start_command/sasl_start_command.dart';
 export 'authentication_commands/sasl_start_command/sasl_start_options.dart';
 
@@ -55,4 +54,3 @@ export 'query_and_write_operation_commands/wrapper/insert_one/insert_one_operati
 export 'query_and_write_operation_commands/wrapper/insert_many/insert_many_operation.dart';
 export 'administration_commands/create_index_operation/create_index_options.dart';
 export 'query_and_write_operation_commands/wrapper/insert_one/insert_one_options.dart';
-export 'query_and_write_operation_commands/find_operation/find_options.dart';